code restructuring for better readibility
