--- pilott/content/__init__.py
@@ -1,1 +1,0 @@
-from pilott.content.content_manager import KnowledgeManager
