+++ pilott/core/__init__.py
@@ -3,8 +3,8 @@ from pilott.core.config import AgentConfig, LLMConfig, LogConfig
 from pilott.core.memory import Memory
 from pilott.core.factory import AgentFactory
 from pilott.core.router import TaskRouter
-from pilott.enums.role import AgentRole
-from pilott.enums.status import AgentStatus
+from pilott.enums.role_e import AgentRole
+from pilott.enums.status_e import AgentStatus
 from pilott.core.task import Task, TaskResult
 
 __all__ = [
