+++ pilott/core/config.py
@@ -1,7 +1,7 @@
 import shutil
 from typing import Optional, List, Dict, Any
 from pydantic import BaseModel, ConfigDict, Field, SecretStr, field_validator
-from pilott.enums.role import AgentRole
+from pilott.enums.role_e import AgentRole
 from pathlib import Path
 from cryptography.fernet import Fernet
 import json
