+++ pilott/enums/__init__.py
@@ -1,6 +1,6 @@
-from pilott.enums.health import HealthStatus
-from pilott.enums.process import ProcessType
-from pilott.enums.role import AgentRole
-from pilott.enums.status import AgentStatus
+from pilott.enums.health_e import HealthStatus
+from pilott.enums.process_e import ProcessType
+from pilott.enums.role_e import AgentRole
+from pilott.enums.status_e import AgentStatus
 from pilott.enums.task_e import TaskPriority, TaskStatus
 from pilott.enums.tool_e import ToolStatus
