+++ pilott/enums/process_e.py
@@ -4,3 +4,5 @@ class ProcessType(str, Enum):
     """Processing type for task execution"""
     SEQUENTIAL = "sequential"
     PARALLEL = "parallel"
+    HIERARCHICAL = "hierarchical"
+    HYBRID = "hybrid"
