+++ pilott/source/__init__.py
@@ -1,0 +1,1 @@
+from pilott.source.source import KnowledgeManager
